omnivex/requirements.txt
yfinance
pandas
requests
beautifulsoup4
jinja2
colorama
psycopg2-binary
python-dotenv
schwab-py
